WT-10246 Use a previous commit when updating a branch fails fab update (#25)
We use a previous commit hash when the current branch does a pull and the build fails. As we cannot undo a pull, and the branch is updated to a newer commit we will use an older commit and check that out in a detached state. This will not create a new branch on the repository, but will act like a branch for the remote framework to use. This can then be updated to a newer branch, or be used if an update to a newer branch fails, which can act as an old_branch.